+++ packages/core/test/unit/stryker-cli.spec.ts
@@ -3,7 +3,6 @@ import sinon from 'sinon';
 import { expect } from 'chai';
 import { DashboardOptions, StrykerOptions, ReportType, PartialStrykerOptions } from '@stryker-mutator/api/core';
 
-import { LogConfigurator } from '../../src/logging/index.js';
 import { guardMinimalNodeVersion, StrykerCli } from '../../src/stryker-cli.js';
 
 describe(StrykerCli.name, () => {
